+++ src/read_resume_data.cpp
@@ -86,7 +86,7 @@ namespace {
 		}
 
 		auto info_hash = rd.dict_find_string_value("info-hash");
-		if (info_hash.size() != sha1_hash::size())
+		if (info_hash.size() != static_cast<std::size_t>(sha1_hash::size()))
 		{
 			ec = errors::missing_info_hash;
 			return ret;
